Move schema validation result parsing code to dedicated file
Additional JSON schema validation result parsing functions will be added so having them in a dedicated file, separate
them from the schema compilation and validation code, will make the code base easier to navigate.
